+++ src/imstat_adios_single_read.cpp
@@ -10,8 +10,8 @@
 // This will work only on 4d images with dimension of polarisation axis 1
 int main(void)
 {
-    int naxis;
-    int naxes[4];
+    int64_t naxis;
+    int64_t naxes[4];
 
     adios2::fstream inStream("casa.bp", adios2::fstream::in_random_access);
 
@@ -29,7 +29,7 @@ int main(void)
     const std::vector<float> data = inStream.read<float>("data");
 
     /* process image one channel at a time; increment channel # in each loop */
-    for (int channel = 0; channel < naxes[2]; channel++)
+    for (int64_t channel = 0; channel < naxes[2]; channel++)
     {
         float sum = 0., meanval = 0., minval = 1.E33, maxval = -1.E33;
         float valid_pix = 0;
@@ -51,7 +51,7 @@ int main(void)
         minval *= 1000.0;
         maxval *= 1000.0;
 
-        printf("%8d %15.6f %10.3f %10.3f %10.3f %10.3f %10.3f %10.3f %10.3f\n",
+        printf("%8lld %15.6f %10.3f %10.3f %10.3f %10.3f %10.3f %10.3f %10.3f\n",
                channel + 1, 1.0f, meanval, 0.0f, 0.0f, 0.0f, 0.0f, minval, maxval);
     }
